+++ src/upload/hooks/useUpload.ts
@@ -7,11 +7,14 @@ import {
   upload,
   getTriggerTextField,
   getDisplayFiles,
+  formatToUploadFile,
 } from '../../_common/js/upload/main';
 import { getFileUrlByFileRaw } from '../../_common/js/upload/utils';
 import useVModel from '../../hooks/useVModel';
 import { InnerProgressContext, OnResponseErrorContext, SuccessContext } from '../../_common/js/upload/types';
 import { useConfig } from '../../hooks/useConfig';
+
+export type ValidateParams = Parameters<TdUploadProps['onValidate']>[0];
 
 /**
  * 上传组件全部逻辑，方便脱离 UI，自定义 UI 组件
@@ -30,6 +33,7 @@ export default function useUpload(props: TdUploadProps) {
 
   const tipsClasses = `${classPrefix.value}-upload__tips ${classPrefix.value}-size-s`;
   const errorClasses = [tipsClasses].concat(`${classPrefix.value}-upload__tips-error`);
+  const placeholderClass = `${classPrefix.value}-upload__placeholder`;
 
   // 单文件场景：触发元素文本
   const triggerUploadText = computed(() => {
@@ -115,19 +119,17 @@ export default function useUpload(props: TdUploadProps) {
       : `${t(locale.value.sizeLimitMessage, { sizeLimit: limit.size })} ${limit.unit}`;
   }
 
-  const handleNonAutoUpload = (toFiles: UploadFile[]) => {
-    // isBatchUpload 场景下，只要没有上传过，就不需要整体替换。只有上传过的文件才需要整体替换
-    const tmpFiles =
-      props.multiple && !(isBatchUpload.value && uploadValue.value[0]?.status === 'success')
-        ? uploadValue.value.concat(toFiles)
-        : toFiles;
+  const handleNotAutoUpload = (toFiles: UploadFile[]) => {
+    const tmpFiles = props.multiple && !isBatchUpload.value ? uploadValue.value.concat(toFiles) : toFiles;
+    if (!tmpFiles.length) return;
+
     // 图片需要本地预览
     if (['image', 'image-flow'].includes(props.theme)) {
       const list = tmpFiles.map(
         (file) =>
           new Promise((resolve) => {
             getFileUrlByFileRaw(file.raw).then((url) => {
-              resolve({ ...file, url });
+              resolve({ ...file, url: file.url || url });
             });
           }),
       );
@@ -135,7 +137,8 @@ export default function useUpload(props: TdUploadProps) {
         setUploadValue(files, {
           trigger: 'add',
           index: uploadValue.value.length,
-          file: files[0],
+          file: toFiles[0],
+          files: toFiles,
         });
       });
     } else {
@@ -150,8 +153,9 @@ export default function useUpload(props: TdUploadProps) {
 
   const onFileChange = (files: FileList) => {
     if (disabled.value) return;
+    const params = { currentSelectedFiles: formatToUploadFile([...files], props.format) };
     // @ts-ignore
-    props.onSelectChange?.([...files], { currentSelectedFiles: toUploadFiles });
+    props.onSelectChange?.([...files], params);
     validateFile({
       uploadValue: uploadValue.value,
       // @ts-ignore
@@ -166,33 +170,48 @@ export default function useUpload(props: TdUploadProps) {
       beforeAllFilesUpload: props.beforeAllFilesUpload,
     }).then((args) => {
       // 自定义全文件校验不通过
-      if (args.validateResult?.type === 'BEFORE_ALL_FILES_UPLOAD') return;
+      if (args.validateResult?.type === 'BEFORE_ALL_FILES_UPLOAD') {
+        const params: ValidateParams = { type: 'BEFORE_ALL_FILES_UPLOAD', files: args.files };
+        props.onValidate?.(params);
+        return;
+      }
       // 文件数量校验不通过
       if (args.lengthOverLimit) {
-        props.onValidate?.({ type: 'FILES_OVER_LENGTH_LIMIT', files: args.files });
+        const params: ValidateParams = { type: 'FILES_OVER_LENGTH_LIMIT', files: args.files };
+        props.onValidate?.(params);
+        if (!args.files.length) return;
       }
       // 过滤相同的文件名
       if (args.hasSameNameFile) {
-        props.onValidate?.({ type: 'FILTER_FILE_SAME_NAME', files: args.files });
+        const params: ValidateParams = { type: 'FILTER_FILE_SAME_NAME', files: args.files };
+        props.onValidate?.(params);
       }
-      // 文件大小校验结果处理
+      // 文件大小校验结果处理（已过滤超出限制的文件）
       if (args.fileValidateList instanceof Array) {
-        const { sizeLimitErrors, toFiles } = getFilesAndErrors(args.fileValidateList, getSizeLimitError);
+        const { sizeLimitErrors, beforeUploadErrorFiles, toFiles } = getFilesAndErrors(
+          args.fileValidateList,
+          getSizeLimitError,
+        );
         const tmpWaitingFiles = autoUpload.value ? toFiles : toUploadFiles.value.concat(toFiles);
         toUploadFiles.value = tmpWaitingFiles;
         props.onWaitingUploadFilesChange?.({ files: tmpWaitingFiles, trigger: 'validate' });
-        // 错误信息处理
+        // 文件大小处理
         if (sizeLimitErrors[0]) {
           sizeOverLimitMessage.value = sizeLimitErrors[0].file.response.error;
           props.onValidate?.({ type: 'FILE_OVER_SIZE_LIMIT', files: sizeLimitErrors.map((t) => t.file) });
         } else {
           sizeOverLimitMessage.value = '';
+          // 自定义方法 beforeUpload 拦截的文件
+          if (beforeUploadErrorFiles.length) {
+            const params: ValidateParams = { type: 'CUSTOM_BEFORE_UPLOAD', files: beforeUploadErrorFiles };
+            props.onValidate?.(params);
+          }
         }
         // 如果是自动上传
         if (autoUpload.value) {
           uploadFiles(tmpWaitingFiles);
         } else {
-          handleNonAutoUpload(tmpWaitingFiles);
+          handleNotAutoUpload(tmpWaitingFiles);
         }
       }
     });
@@ -201,7 +220,7 @@ export default function useUpload(props: TdUploadProps) {
     inputRef.value.value = null;
   };
 
-  const onNormalFileChange = (e: Event) => {
+  const onNormalFileChange = (e: InputEvent) => {
     onFileChange?.((e.target as HTMLInputElement).files);
   };
 
@@ -210,8 +229,8 @@ export default function useUpload(props: TdUploadProps) {
   }
 
   /**
-   * 上传文件
-   * 对外暴露方法，修改时需谨慎
+   * 上传文件。对外暴露方法，修改时需谨慎
+   * @param toFiles 本地上传的文件列表
    */
   function uploadFiles(toFiles?: UploadFile[]) {
     const notUploadedFiles = uploadValue.value.filter((t) => t.status !== 'success');
@@ -232,6 +251,7 @@ export default function useUpload(props: TdUploadProps) {
       uploadAllFilesInOneRequest: props.uploadAllFilesInOneRequest,
       useMockProgress: props.useMockProgress,
       data: props.data,
+      mockProgressDuration: props.mockProgressDuration,
       requestMethod: props.requestMethod,
       formatRequest: props.formatRequest,
       formatResponse: props.formatResponse,
@@ -253,6 +273,7 @@ export default function useUpload(props: TdUploadProps) {
               file: data.files[0],
             });
           }
+          xhrReq.value = [];
           props.onSuccess?.({
             fileList: data.files,
             currentFiles: files,
@@ -261,8 +282,8 @@ export default function useUpload(props: TdUploadProps) {
             results: list?.map((t) => t.data),
             // 单文件单请求有一个 response，多文件多请求有多个 response
             response: data.response || list.map((t) => t.data.response),
+            XMLHttpRequest: data.XMLHttpRequest,
           });
-          xhrReq.value = [];
         } else if (failedFiles?.[0]) {
           props.onFail?.({
             e: data.event,
@@ -270,6 +291,7 @@ export default function useUpload(props: TdUploadProps) {
             failedFiles,
             currentFiles: files,
             response: data.response,
+            XMLHttpRequest: data.XMLHttpRequest,
           });
         }
 
@@ -278,10 +300,6 @@ export default function useUpload(props: TdUploadProps) {
           toUploadFiles.value = failedFiles;
           props.onWaitingUploadFilesChange?.({ files: failedFiles, trigger: 'uploaded' });
         }
-      },
-      (p) => {
-        onResponseError(p);
-        uploading.value = false;
       },
     );
   }
@@ -304,6 +322,7 @@ export default function useUpload(props: TdUploadProps) {
       uploadValue.value.splice(p.index, 1);
       setUploadValue([...uploadValue.value], changePrams);
     } else {
+      // autoUpload 场景下， p.index < uploadValue.length 表示移除已经上传成功的文件；反之表示移除待上传列表文件
       // eslint-disable-next-line
       if (p.index < uploadValue.value.length) {
         uploadValue.value.splice(p.index, 1);
@@ -318,7 +337,7 @@ export default function useUpload(props: TdUploadProps) {
   }
 
   const triggerUpload = (e?: MouseEvent) => {
-    if (disabled.value) return;
+    if (disabled.value || !inputRef.value) return;
     e?.stopPropagation?.();
     (inputRef.value as HTMLInputElement).click();
   };
@@ -343,9 +362,11 @@ export default function useUpload(props: TdUploadProps) {
       );
     }
 
-    if (context?.file) {
+    if (context?.file && !autoUpload.value) {
       onInnerRemove?.({ file: context.file, e: context.e, index: 0 });
     }
+
+    props.onCancelUpload?.();
   };
 
   return {
@@ -360,6 +381,7 @@ export default function useUpload(props: TdUploadProps) {
     uploading,
     tipsClasses,
     errorClasses,
+    placeholderClass,
     inputRef,
     disabled,
     xhrReq,